Move excon integration tests to the appropriate file
