+++ spec/integration/excon_spec.rb
@@ -79,6 +79,21 @@ describe HTTPI::Adapter::HTTPClient do
       expect(response.body).to eq("basic-auth")
     end
 
+    it "does not support ntlm authentication" do
+      request = HTTPI::Request.new(@server.url + "ntlm-auth")
+      request.auth.ntlm("tester", "vReqSoafRe5O")
+
+      expect { HTTPI.get(request, adapter) }.
+        to raise_error(HTTPI::NotSupportedError, /does not support NTLM authentication/)
+    end
+
+    it "supports disabling verify mode" do
+      request = HTTPI::Request.new(@server.url)
+      request.auth.ssl.verify_mode = :none
+      adapter_class = HTTPI::Adapter.load(adapter).new(request)
+      expect(adapter_class.client.data[:ssl_verify_peer]).to eq(false)
+    end
+
     it "supports chunked response" do
       request = HTTPI::Request.new(@server.url)
       res = ""
@@ -109,6 +124,27 @@ describe HTTPI::Adapter::HTTPClient do
       it "works when set up properly" do
         request = HTTPI::Request.new(@server.url)
         request.auth.ssl.ca_cert_file = IntegrationServer.ssl_ca_file
+
+        response = HTTPI.get(request, adapter)
+        expect(response.body).to eq("get")
+      end
+
+      it "works with client cert and key provided as file path" do
+        request = HTTPI::Request.new(@server.url)
+        request.auth.ssl.ca_cert_file = IntegrationServer.ssl_ca_file
+        request.auth.ssl.cert_file = "spec/fixtures/client_cert.pem"
+        request.auth.ssl.cert_key_file = "spec/fixtures/client_key.pem"
+
+        response = HTTPI.get(request, adapter)
+        expect(response.body).to eq("get")
+      end
+
+      it "works with client cert and key set directly" do
+        request = HTTPI::Request.new(@server.url)
+
+        request.auth.ssl.ca_cert_file = IntegrationServer.ssl_ca_file
+        request.auth.ssl.cert = OpenSSL::X509::Certificate.new File.open("spec/fixtures/client_cert.pem").read
+        request.auth.ssl.cert_key = OpenSSL::PKey.read File.open("spec/fixtures/client_key.pem").read
 
         response = HTTPI.get(request, adapter)
         expect(response.body).to eq("get")
